+++ docs/scripts/config/menu.js
@@ -58,10 +58,10 @@ const MATERIAL_COMPONENTS_PRO = [
       },
       {
         name: 'constant'
-      },
-      {
-        name: 'dialog'
       }
+      // {
+      //   name: 'dialog'
+      // }
     ]
   },
   '-',